Add a structure for changes in the refenrece set.
Until now all changes to reference sets were simply stored as a single
string (for all elements within the set) in a LogLine, making it hard to
search and view the changes. With these changes we store the fields of
each element in the set in separate ReferenceSetLog and
ReferenceSetLogLines which can then be visualized in a structured
manner.