Fix contact person triggers on create and delete
It had two triggers on create, one before populating the data and one after.
This goes against what the documentation recommends, which is to follow what
has been done to the audit trail super class.

Also, on delete, logs should not be deleted (through the ContactPerson_Log
association).

This should be fixed in the documentation as well.